pyproject.toml
[tool.poetry]
name = "dropmate-py"
version = "0.1.0"
description = "Python helpers for the EDC Dropmate"
authors = ["sco1 <sco1.git@gmail.com>"]

readme = "README.md"
homepage = "https://github.com/sco1/"
repository = "https://github.com/sco1/dropmate-py"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

include = [
    {path = "./CHANGELOG.md", format = "sdist"}
]

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/sco1/dropmate-py/issues"

[tool.poetry.scripts]
dropmate = "dropmate_py.cli:dropmate_cli"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0"
sco1-misc = "^0.1"
typer = "^0.9"

[tool.poetry.dev-dependencies]
black = "^23.1"
bump2version = "^1.0"
cogapp = "^3.3"
flake8 = "^6.0"
flake8-annotations = "^3.0"
flake8-bugbear = "^23.1"
flake8-docstrings = "^1.7"
flake8-fixme = "^1.1"
isort = "^5.12"
mypy = "^1.0"
pep8-naming = "^0.13"
pre-commit = "^3.0"
pytest = "^7.2"
pytest-cov = "^4.0"
pytest-randomly = "^3.12"
tox = "^4.4"

[tool.black]
line-length = 100

[tool.isort]
case_sensitive = true
known_first_party = "dropmate_py,tests"
no_lines_before = "LOCALFOLDER"
order_by_type = false
profile = "black"
line_length = 100

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core>=1.2"]
build-backend = "poetry.core.masonry.api"
